the-master-os/apps/api/pyproject.toml
[project]
name = "the-master-os-api"
version = "0.1.0"
description = "The Master OS — Orchestration Engine (FastAPI + LangGraph)"
requires-python = ">=3.12"
readme = {text = "The Master OS Orchestration Engine API", content-type = "text/plain"}
license = { text = "Proprietary" }

dependencies = [
    # Core Framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",

    # Database & Storage
    "supabase>=2.11",

    # HTTP & Network
    "httpx>=0.28",
    "tenacity>=9.0",

    # Security
    "python-jose[cryptography]>=3.3",
    "cryptography>=44.0",
    "python-multipart>=0.0.18",

    # Task Queue
    "celery[redis]>=5.4",
    "redis>=5.2",

    # Rate Limiting
    "slowapi>=0.1.9",

    # Monitoring
    "sentry-sdk[fastapi]>=2.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "mypy>=1.13",
    "ruff>=0.8",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "S",    # flake8-bandit (security)
]
ignore = ["S101"]  # allow assert in tests

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
